rename module to `on_disk_numeric_index`
