+++ lib/segment/src/index/field_index/numeric_index/builders.rs
@@ -9,8 +9,8 @@ use gridstore::Blob;
 use serde_json::Value;
 
 use super::mutable_numeric_index::InMemoryNumericIndex;
+use super::on_disk_numeric_index::OnDiskNumericIndex;
 use super::storage::NumericIndexInner;
-use super::universal_numeric_index::OnDiskNumericIndex;
 use super::{Encodable, NumericIndex, NumericIndexIntoInnerValue};
 use crate::common::operation_error::{OperationError, OperationResult};
 use crate::index::field_index::numeric_point::Numericable;
