+++ lib/segment/src/index/field_index/numeric_index/on_disk_numeric_index/lifecycle.rs
