+++ lib/segment/src/index/field_index/numeric_index/storage/mod.rs
@@ -31,7 +31,7 @@ use gridstore::Blob;
 use super::Encodable;
 use super::immutable_numeric_index::ImmutableNumericIndex;
 use super::mutable_numeric_index::MutableNumericIndex;
-use super::universal_numeric_index::OnDiskNumericIndex;
+use super::on_disk_numeric_index::OnDiskNumericIndex;
 use crate::index::field_index::numeric_point::Numericable;
 use crate::index::field_index::stored_point_to_values::StoredValue;
