+++ lib/segment/src/index/field_index/numeric_index/storage/read_only/lifecycle.rs
@@ -8,7 +8,7 @@ use super::super::super::Encodable;
 use super::super::super::mutable_numeric_index::read_only::ReadOnlyAppendableNumericIndex;
 use super::ReadOnlyNumericIndexInner;
 use crate::common::operation_error::OperationResult;
-use crate::index::field_index::numeric_index::universal_numeric_index::OnDiskNumericIndex;
+use crate::index::field_index::numeric_index::on_disk_numeric_index::OnDiskNumericIndex;
 use crate::index::field_index::numeric_point::Numericable;
 use crate::index::field_index::stored_point_to_values::StoredValue;
 use crate::index::payload_config::IndexMutability;
